% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRI.R
\name{calRI}
\alias{calRI}
\title{Calculate Rand Index}
\usage{
calRI(a, b)
}
\arguments{
\item{a}{A vector that contains the cluster label of each sample.}

\item{b}{A vector that contains the true label of each sample.}
}
\value{
A real number that means the rand index between the two clustering results.
}
\description{
Calculate the Rand Index of the clustering results.
}
\examples{
a=c(1,3,2,2,4,6,5)
b=c(3,1,2,5,6,4,3)
calRI(a,b)
}
