% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPICS.R
\name{CEPICS}
\alias{CEPICS}
\title{CEPICS: a comparison and evaluation platform for integration methods in cancer subtyping}
\usage{
CEPICS(datalist, datatype, functionList = list("PINS", "LRA", "SNF",
  "iCluster"), kMax = 5, Res = NULL, KMdata = "default",
  trueLabel = NULL, SNFPara = NULL, LRAPara = NULL,
  iClusterPara = NULL, PINSPara = NULL, filename = "CEPICS",
  title = "Evaluation and Comparison Report", author = "", cores = 0)
}
\arguments{
\item{datalist}{A list of different omics data. Each data should be formatted as a data matrix with rows representing features and columns representing  samples.}

\item{datatype}{A list of string means the corresponding datatype of datalist. Each type can choose from "binary", "gaussian" or "poisson".}

\item{functionList}{A list of functions that need to be ran, choosing from "iCluster" (means iClusterBayes), "SNF", "LRA" (means LRAcluster), "PFA", and "PINS". We set SNF, LRA, iClusterBayes and PINS as default.}

\item{kMax}{An integer value means the maximize number of clusters we will try.Res    A list contains the result of user's method. Please see Res for more details.}

\item{Res}{User's result.Put your data into \code{Res$data}, and put name into \code{Res$name}, put the time you cost into \code{Res$time}. If this parameter is not null, this data will appear at the top of the report.}

\item{KMdata}{The survival data of the sample. If this parameter is "default", CEPICS will try to use built-in data including BRCA, COAD, KIRC, LUAD and LUSC downloaded from TCGA. If this parameter is FALSE, CEPICS will not to calculate the KM curve.}

\item{trueLabel}{A list of the true labels of the samples. If this parameter is null, CEPICS will compare different metrices between each two methods. Otherwise the results of each methods will be compared with the true labels. Please see turelabel for more details.}

\item{SNFPara}{Parameters of SNF. Should be a list of following parameters, \code{SNFPara$k}(default is one tenth of the sample size), \code{SNFPara$t}(default is 20). Please see runSNF for more details. If the parameter is null, CEPICS will use default values.}

\item{LRAPara}{Parameters of LRA. Should be a list of following parameter, \code{LRAPara$maxdimension} (default is 10) is the max number of dimension CEPICS will try to reduce. If the parameter is null, CEPICS will use default values.}

\item{iClusterPara}{Parameters of iClusterBayes. Should be a list of following parameters, \code{iClusterPara$n.burnin}(default is 1000), \code{iClusterPara$n.draw}(default is 1200), \code{iClusterPara$prior.gamma}(default is rep(0.1,6)), \code{iClusterPara$sdev}(default is 0.5), \code{iClusterPara$beta.var.scale}(default is 1), \code{iClusterPara$thin}(default is 1), \code{iClusterPara$pp.cutoff}(default is 0.5). Please see iClusterPlus R package for more details. If the parameter is null, CEPICS will use default values.}

\item{PINSPara}{Parameters of PINS. Should be a list of following parameter, \code{PINSPara$agreementCutoff}(default is 0.5) means agreement threshold to be considered consistent. If the parameter is null, CEPICS will use default values.}

\item{filename}{The file name of the report.}

\item{title}{The title of the report.}

\item{author}{The author of thereport.}

\item{cores}{An integer value means the number of cores for parallel computing, default 0. When using the default value, CEPICS will detect the physical cores and set cores-1 as the number of cores to parallel compute.}
}
\value{
Output the evaluation and comparion report and some experimental data files.
}
\description{
The main function of CEPICS R package Please refer to examples for details. When running CEPICS, you don't need to initialize the MATLAB server since  CEPICS will initialize it automatically.
}
\examples{
data(COAD_Methy)
data(COAD_miRNA)
data(COAD_mRNA)


datalist <- list(COAD_Methy,COAD_miRNA,COAD_mRNA)

\donttest{
CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'), kMax=5)

#Use Survival data
data(KMSURDATA)
CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'), kMax=5, KMdata = KMSURDATA)

#Use your result
data(Res)

CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'), kMax=5, Res = Res)

#Use gold standard
data(truelabel)
CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'), kMax=5, trueLabel = truelabel)

#Pass parameter to function
CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    kMax=5, SNFPara = list(k = 8, t = 20))

CEPICS(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    kMax=5, LRAPara = list(maxdimension=10))
}


}
