% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPIM.R
\name{CEPIM}
\alias{CEPIM}
\title{CEPIM, a comparison and evaluation platform for integration methods in cancer subtyping}
\usage{
CEPIM(datalist, datatype, functionList = list("PINS", "LRA", "SNF",
  "iCluster"), evalList = list("Heatmap", "KM", "SI", "ARI", "RI",
  "NMI"), kMax = 5, Res = NULL, KMdata = NULL, trueLabel = NULL,
  SNFPara = NULL, LRAPara = NULL, iClusterPara = NULL,
  PINSPara = NULL, filename = "CEPIM",
  title = "Evaluation and Comparison Report", author = "")
}
\arguments{
\item{datalist}{A list of different omics data. Each data should be formatted as a data matrix with rows representing features and columns representing  samples.}

\item{datatype}{A list of string means the corresponding datatype of datalist. Each type can choose from "binary", "gaussian" or "poisson".}

\item{functionList}{A list of functions that need to be ran, choosing from "iCluster" (means iClusterBayes), "SNF", "LRA" (means LRAcluster), "PFA", and "PINS". We set SNF, LRA, iClusterBayes and PINS as default.evalList    The evaluation metrices that needs to be evaluated and compared, choosing from "KM", "SI", "Heatmap", "RI", "ARI", and "NMI". We set all the metrices as default.}

\item{evalList}{The evaluation method that needs to be ran. Should be some of "KM", "SI", "Heatmap", "RI", "ARI", "NMI".}

\item{kMax}{An integer value means the maximize number of clusters we will try.Res    A list contains the result of user's method. Please see Res for more details.}

\item{Res}{User's result.Put your data into \code{Res$data}, and put name into \code{Res$name}, put the time you cost into \code{Res$time}. If this parameter is not null, this data will appear at the top of the report.}

\item{KMdata}{The survival data of the sample. If this parameter is null, CEPIM will try to use built-in data including BRCA, COAD, KIRC, LUAD and LUSC downloaded from TCGA.}

\item{trueLabel}{A list of the true labels of the samples. If this parameter is null, CEPIM will compare different metrices between each two methods. Otherwise the results of each methods will be compared with the true labels. Please see turelabel for more details.}

\item{SNFPara}{Parameters of SNF. Should be a list of following parameters, \code{SNFPara$k}(default is one tenth of the sample size), \code{SNFPara$t}(default is 20). Please see runSNF for more details. If the parameter is null, CEPIM will use default values.}

\item{LRAPara}{Parameters of LRA. Should be a list of following parameter, \code{LRAPara$maxdimension} (default is 10) is the max number of dimension CEPIM will try to reduce. If the parameter is null, CEPIM will use default values.}

\item{iClusterPara}{Parameters of iClusterBayes. Should be a list of following parameters, \code{iClusterPara$n.burnin}(default is 1000), \code{iClusterPara$n.draw}(default is 1200), \code{iClusterPara$prior.gamma}(default is rep(0.1,6)), \code{iClusterPara$sdev}(default is 0.5), \code{iClusterPara$beta.var.scale}(default is 1), \code{iClusterPara$thin}(default is 1), \code{iClusterPara$pp.cutoff}(default is 0.5). Please see iClusterPlus R package for more details. If the parameter is null, CEPIM will use default values.}

\item{PINSPara}{Parameters of PINS. Should be a list of following parameter, \code{PINSPara$agreementCutoff}(default is 0.5) means agreement threshold to be considered consistent. If the parameter is null, CEPIM will use default values.}

\item{filename}{The file name of the report.}

\item{title}{The title of the report.}

\item{author}{The author of thereport.}
}
\value{
Output the evaluation and comparion report and some experimental data files.
}
\description{
The main function of CEPIM R package Please refer to examples for details. When running CEPIM, you don't need to initialize the MATLAB server since  CEPIM will initialize it automatically.
}
\examples{
data(COAD_Methy)
data(COAD_miRNA)
data(COAD_mRNA)


datalist <- list(COAD_Methy,COAD_miRNA,COAD_mRNA)

\donttest{
CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'), kMax=5)

#Use Survival data
data(KMSURDATA)
CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'), kMax=5, KMdata = KMSURDATA)

#Use your result
data(Res)

CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'), kMax=5, Res = Res)

#Use gold standard
data(truelabel)
CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'), kMax=5, trueLabel = truelabel)

#Pass parameter to function
CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'),
    kMax=5, SNFPara = list(k = 8, t = 20))

CEPIM(datalist, datatype = c("gaussian", "gaussian", "gaussian"),
    functionList = list('PINS', 'LRA', 'SNF', 'iCluster', 'PFA'),
    evalList = list('Heatmap','KM', 'SI', 'ARI', 'RI', 'NMI'),
    kMax=5, LRAPara = list(maxdimension=10))
}


}
