% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAD_cnv.R
\docType{data}
\name{COAD_cnv}
\alias{COAD_cnv}
\title{A Copy Number Variations dataset of COAD.}
\format{A data matrix with 1961 rows and 255 columns.}
\usage{
COAD_cnv
}
\description{
A Colon adenocarcinoma (COAD) copy number variations dataset downloaded from TCGA and pre-processed using CEPICS.
}
\examples{
data(COAD_cnv)
}
\keyword{datasets}
