% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAD_Methy.R
\docType{data}
\name{COAD_Methy}
\alias{COAD_Methy}
\title{A DNA methylation dataset of COAD.}
\format{A data matrix with 2300 rows and 255 columns.}
\usage{
COAD_Methy
}
\description{
A Colon adenocarcinoma (COAD) DNA methylation dataset pre-processed using CEPIM.
}
\examples{
data(COAD_Methy)
}
\keyword{datasets}
