% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAD_miRNA.R
\docType{data}
\name{COAD_miRNA}
\alias{COAD_miRNA}
\title{A microRNA dataset of COAD}
\format{A data matrix with 78 rows and 255 columns.}
\usage{
COAD_miRNA
}
\description{
A Colon adenocarcinoma (COAD) microRNA dataset downloaded from TCGA and pre-processed using CEPICS.
}
\examples{
data(COAD_miRNA)
}
\keyword{datasets}
