% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAD_mRNA.R
\docType{data}
\name{COAD_mRNA}
\alias{COAD_mRNA}
\title{A message RNA dataset of COAD}
\format{A data matrix with 2071 rows and 255 columns.}
\usage{
COAD_mRNA
}
\description{
A Colon adenocarcinoma (COAD) message RNA dataset downloaded from TCGA and pre-processed using CEPICS.
}
\examples{
data(COAD_mRNA)
}
\keyword{datasets}
