% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter Data}
\usage{
filterData(data, alpha1 = 1, alpha2 = 1, filter_zero_as_NA = FALSE)
}
\arguments{
\item{data}{The data needs to be filter.}

\item{alpha1}{The ratio of NA values in a column is greater than alpha1 and the column will be discarded.}

\item{alpha2}{The ratio of NA values in a row is greater than alpha1 and the row  will be discarded.}

\item{filter_zero_as_NA}{If this parameter is true, the 0 in the data will be treated as NA.}
}
\value{
Filtered data.
}
\description{
Discard rows or columns whose NA value is greater than a certain percentage.
}
\details{
Normally, alpha1 and alpha2 only need to use one of them.
}
\examples{
data(COAD_Methy)
COAD_Methy <- filterData(COAD_Methy, alpha1 = 0.8)

}
