% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Format Data}
\usage{
formatData(data, trans = FALSE, sep = ",")
}
\arguments{
\item{data}{File path of the data that needs to be read.}

\item{trans}{Whether you need to transpose the results.}

\item{sep}{The separator of the data in the file.}
}
\value{
Data in matrix format read from file.
}
\description{
Read data from a file.The first row and the first column are the names of the data.
}
\examples{
data(COAD_Methy)
write.csv(COAD_Methy, 'temp.csv')
mydata <- formatData('temp.csv', sep = ',')

}
