% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlap.R
\name{getOverlap}
\alias{getOverlap}
\title{Get the overlap samples of  datasets.}
\usage{
getOverlap(datalist)
}
\arguments{
\item{datalist}{A list object. Different omics datasets should be gathered as a datalist.}
}
\description{
Select samples from different omics datasets that belong to the same patient cohort.
}
\examples{
data(COAD_Methy, COAD_miRNA, COAD_mRNA)
datalist <- list(COAD_Methy, COAD_miRNA, COAD_mRNA)
datalist <- getOverlap(datalist)
}
