% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Impute the missing values of Data.}
\usage{
imputeData(data, method = "knn", k = 10, trans_zero_to_NA = FALSE)
}
\arguments{
\item{data}{The data matrix needs to be imputed.}

\item{method}{Choose from "knn", "mean", and "median". "knn": using K nearest neighbors methods to impute the missing value.
"mean": using the mean value of each feature to impute the missing value.
"median": using the median value of each feature to impute the missing value.}

\item{k}{An integer value which means the number of the nearest neighbors when using KNN method to impute missing values.}

\item{trans_zero_to_NA}{A logical value. If TRUE, all zeros in the data will be converted to NA.}
}
\description{
Using different methods to impute the missing values of the data. CEPIM provides three imputation methods.
}
\examples{
data(COAD_Methy)
COAD_Methy <- imputeData(COAD_Methy, method = 'mean')
}
\references{
Troyanskaya,O. et al. (2001) Missing value estimation methods for DNA microarrays. Bioinformatics, 17, 520-525.
}
