% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSURDATA.R
\docType{data}
\name{KMSURDATA}
\alias{KMSURDATA}
\title{Clinical Survival Data of five TCGA cancers.}
\format{An object of class dataframe with 3119 rows and 3 columns.}
\usage{
KMSURDATA
}
\description{
In order to generate Kaplan-Meier survival plots and calculate P-values, clinical survival data should be provided as input. For convenience, CEPICS collected clinical data of five cancers (BRCA, COAD, KIRC, LUAD, and LUSC) of TCGA. This dataset gives the clinical survival data of 3119 patients.
}
\examples{
data(KMSURDATA)
}
\keyword{datasets}
