% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normData.R
\name{normData}
\alias{normData}
\title{Normalize Data}
\usage{
normData(data, method = "z_score", object = "feature", logBase = 2)
}
\arguments{
\item{data}{A data matrix which needs to be normalized.}

\item{method}{A string means the method of normalization, choose from "log", "z_score", "mean", "median".
"log": using log transformation to normalize the data, which means take the logarithm of each value to base X. X should be set as parameter "logBase".
"z-score": using z-score to normalize the data.
"mean": each value minus the mean value of its feature or sample.
"median": each value minus the median value of its feature or sample.}

\item{object}{Choose of "feature" and "sample" which means normalize for each feature or sample respectively.}

\item{logBase}{The base of log when using log transformation to normalize data.}
}
\description{
Using different methods to normalize the data. CEPICS provides four normalization methods.
}
\examples{
data(COAD_Methy)
COAD_Methy <- normData(COAD_Methy, method = 'z_score', object = 'feature')
}
