% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFAinit.R
\name{PFAinit}
\alias{PFAinit}
\title{Initialize a MATLAB server}
\usage{
PFAinit(path = getwd())
}
\arguments{
\item{path}{A string, which means the work dictionary of MATLAB server. Default is the current R working directory getting from getwd().}
}
\value{
If open MATLAB server successfully, return the access of MATLAB, the variable will be used as the parameter of runPFA.When you have done all of your work, you must close the access by \code{close(matlab)}.
}
\description{
Initialize the MATLAB server on the local computer, add the path as MATLAB workspace, and create PFA source file for MATLAB on the path.
}
\details{
You must have installed MATLAB on your computer.
}
\examples{
\donttest{
matlab=PFAinit(path=getwd())
#now, we can run PFA on MATLAB server

#TODO
#for example:
res=runPFA(data=list(COAD_Methy, COAD_miRNA, COAD_mRNA),maxk=10,matlab=matlab)

#after working on MATLAB, we should close the MATLAB server.
close(matlab)
}

}
\seealso{
\code{\link{runPFA}}
}
