% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiClusterBayes.R
\name{runiClusterBayes}
\alias{runiClusterBayes}
\title{Run the iClusterBayes method}
\usage{
runiClusterBayes(dl, type = c("gaussian", "gaussian", "gaussian",
  "gaussian", "gaussian", "gaussian"), kMax = 4, n.burnin = 1000,
  n.draw = 1200, prior.gamma = rep(0.1, 6), sdev = 0.5,
  beta.var.scale = 1, thin = 1, pp.cutoff = 0.5, cores = 1)
}
\arguments{
\item{dl}{The datalist needs to be processed.}

\item{type}{Data type corresponding to dl1-6, which can be gaussian, binomial, or poisson.}

\item{kMax}{The maximum number of clusters.}

\item{n.burnin}{Please see "iClusterPlus" package for more details.}

\item{n.draw}{Please see "iClusterPlus" package for more details.}

\item{prior.gamma}{Please see "iClusterPlus" package for more details.}

\item{sdev}{Please see "iClusterPlus" package for more details.}

\item{beta.var.scale}{Please see "iClusterPlus" package for more details.}

\item{thin}{Please see "iClusterPlus" package for more details.}

\item{pp.cutoff}{Please see "iClusterPlus" package for more details.}

\item{cores}{An integer value means the number of cores for parallel computing.}
}
\value{
2 to kMax clustering results and sample similarity matrix.
}
\description{
Run the iClusterBayes method
}
\examples{
\donttest{
data(COAD_Methy, COAD_miRNA, COAD_mRNA)
datalist <- list(COAD_Methy, COAD_miRNA, COAD_mRNA)
res <- runiClusterBayes(datalist, type = c("gaussian","gaussian","gaussian"),kMax=5,n.burnin=1000,
     n.draw=1200,prior.gamma=rep(0.1,6),sdev=0.5,beta.var.scale=1,thin=1,pp.cutoff=0.5)
}

}
\references{
Mo,Q. et al. (2017) A fully Bayesian latent variable model for integrative clustering analysis of multi-type omics data. Biostatistics, 19, 71-86.
}
