% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPFA.R
\name{runPFA}
\alias{runPFA}
\title{Run PFA method to get the subtyping results.}
\usage{
runPFA(data, maxk = 10, matlab)
}
\arguments{
\item{data}{A list of different omics data. Each data in data list should be format as a data matrix with rows representing features and columns representing samples.}

\item{maxk}{An integer value means the maximize number of clusters we will try.}

\item{matlab}{MATLAB server variable when execute \code{PFAinit()} function.}
}
\description{
Using PFA to apply dimension reduction and data integration, then using k-means to get the cluster results.
}
\details{
The author just provides the MATLAB code of the method. When we tried to transfer it to R, we found that the results of left division operation for non-square matrices in R and MATLAB were totally different. In order to compare and evaluate the original version of PFA, CEPIM uses R.matlab R package to initialize a MATLAB server to run the original MATLAB code. In addition, the original version code can only integrate three omics datasets. We improve the code and make it possible to integrate omics data without the number of datasets limit. You should execute \code{PFAinit()} to initialize a MATLAB server before running PFA, and close server after running.
}
\examples{
data(COAD_Methy)
data(COAD_miRNA)
data(COAD_mRNA)

\donttest{
matlab=PFAinit(path=getwd())
res=runPFA(data=list(COAD_Methy, COAD_miRNA, COAD_mRNA),maxk=10,matlab=matlab)
close(matlab)
}

}
\references{
Shi,Q. et al. (2017) Pattern fusion analysis by adaptive alignment of multiple heterogeneous omics data. Bioinformatics, 33, 2706-2714.
}
\seealso{
\code{\link{PFAinit}}
}
