% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPINSPlus.R
\name{runPINSPlus}
\alias{runPINSPlus}
\title{Run PINSPlus method to get the subtyping results.}
\usage{
runPINSPlus(datalist, kMax = 5, agreementCutoff = 0.5, cores = 1)
}
\arguments{
\item{datalist}{A list of different omics data. Each data in data list should be format as a data matrix with rows representing features and columns representing samples.}

\item{kMax}{An integer value means the maximize number of clusters we will try.}

\item{agreementCutoff}{agreement threshold to be considered consistent. Default value is 0.5. Please see PINSPlus package for more details.}

\item{cores}{An integer value means the number of cores for parallel computing.}
}
\value{
The sample clustering results matrix.
}
\description{
Using PINSPlus R package to get the subtyping results.
}
\examples{
data(COAD_Methy, COAD_miRNA, COAD_mRNA)
datalist <- list(COAD_Methy, COAD_miRNA, COAD_mRNA)
res <- runPINSPlus(datalist, 5, 0.5)

}
\references{
Nguyen,T. et al. (2017) A novel approach for data integration and disease subtyping. Genome Res., 27, 2025-2039.
}
