% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSNF.R
\name{runSNF}
\alias{runSNF}
\title{Run SNF method to get the subtyping results.}
\usage{
runSNF(Wall, k = NULL, t = 20, kMax = 5)
}
\arguments{
\item{Wall}{A list of different omics data. Each data in data list should be format as a data matrix with rows representing features and columns representing samples.}

\item{k}{An integer number means the number of neighbors in K-nearest neighbors.Default \code{NULL}.}

\item{t}{An integer number means the number of iterations for the diffusion process. Default \code{20}.}

\item{kMax}{An integer value means the maximize number of clusters we will try. Default \code{5}.}
}
\value{
The sample clustering results matrix including 2 to kMax clustering results, and sample similarity matrix.
}
\description{
Using SNFTool R package to get the subtyping results.
}
\examples{
data(COAD_Methy, COAD_miRNA, COAD_mRNA)
datalist <- list(COAD_Methy, COAD_miRNA, COAD_mRNA)
res <- runSNF(datalist, k = 10, t = 20, kMax = 5)

}
\references{
Wang,B. et al. (2014) Similarity network fusion for aggregating data types on a genomic scale. Nature Methods, 11, 333-337.

Concise description can be found here: http://compbio.cs.toronto.edu/SNF/SNF/Software.html
}
