% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeature.R
\name{selectFeature}
\alias{selectFeature}
\title{Select features of dataset}
\usage{
selectFeature(data, method = "VAR", select_number = 1)
}
\arguments{
\item{data}{A data matrix.}

\item{method}{A string means  the method of feature selection, choosing from "MAD", "VAR" and "PCA".
"MAD": using median absolute deviation to select features.
"VAR": select features according to the variance of each feature.
"PCA": using principal component analysis to select features.}

\item{select_number}{A real value from 0 to 1 which means the ratio of features we want to select. For "PCA", this parameter indicates the number of features to be selected instead of the ratio.}
}
\value{
The data matrix after feature selection.
}
\description{
Using different strategies to select features of data.
}
\examples{
data(COAD_Methy)
COAD_Methy <- selectFeature(COAD_Methy, method = 'VAR', select_number = 0.7)

}
