% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truelabel.R
\docType{data}
\name{truelabel}
\alias{truelabel}
\title{the truelabels of COAD samples}
\format{An object of class integer of length 255.}
\usage{
truelabel
}
\description{
An empirical pre-determination of subtypes for samples by experts or clinicians based on clinical phenotypes, images or experience can be offered to CEPIM, which can be used as the gold standard to evaluate and compare results generated by different methods.
A previous study (Sanchez-Vega, et al., 2018) shows three subtypes of COAD defined by the individual TCGA analysis working groups. We take it as true labels when we use COAD datasets of TCGA as the examples of function CEPIM.
}
\examples{
data(truelabel)

}
\references{
Sanchez-Vega,F. et al. (2018) Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173, 321-337.e10.
}
\seealso{
\code{\link{CEPIM}}
}
\keyword{datasets}
